service spun up with deps in k8s with skaffold
