+++ backend/cmd/functions/outbound_emailer/function.go
@@ -7,22 +7,17 @@ import (
 	"log/slog"
 	"os"
 	"strings"
-	"time"
 
-	analyticsconfig "github.com/dinnerdonebetter/backend/internal/analytics/config"
+	"github.com/dinnerdonebetter/backend/cmd/functions/outbound_emailer/logic"
 	"github.com/dinnerdonebetter/backend/internal/config"
-	"github.com/dinnerdonebetter/backend/internal/database/postgres"
 	"github.com/dinnerdonebetter/backend/internal/email"
-	emailconfig "github.com/dinnerdonebetter/backend/internal/email/config"
 	"github.com/dinnerdonebetter/backend/internal/observability"
 	"github.com/dinnerdonebetter/backend/internal/observability/logging"
 	loggingcfg "github.com/dinnerdonebetter/backend/internal/observability/logging/config"
-	"github.com/dinnerdonebetter/backend/internal/observability/tracing"
 
 	_ "github.com/GoogleCloudPlatform/functions-framework-go/funcframework"
 	"github.com/GoogleCloudPlatform/functions-framework-go/functions"
 	"github.com/cloudevents/sdk-go/v2/event"
-	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
 	"go.opentelemetry.io/otel"
 	_ "go.uber.org/automaxprocs"
 )
@@ -55,11 +50,6 @@ func SendEmail(ctx context.Context, e event.Event) error {
 
 	logger := (&loggingcfg.Config{Level: logging.DebugLevel, Provider: loggingcfg.ProviderSlog}).ProvideLogger()
 
-	envCfg := email.GetConfigForEnvironment(os.Getenv("DINNER_DONE_BETTER_SERVICE_ENVIRONMENT"))
-	if envCfg == nil {
-		return observability.PrepareAndLogError(email.ErrMissingEnvCfg, logger, nil, "getting environment config")
-	}
-
 	cfg, err := config.GetOutboundEmailerConfigFromGoogleCloudSecretManager(ctx)
 	if err != nil {
 		return fmt.Errorf("error getting config: %w", err)
@@ -71,31 +61,6 @@ func SendEmail(ctx context.Context, e event.Event) error {
 	}
 	otel.SetTracerProvider(tracerProvider)
 
-	ctx, span := tracing.NewTracer(tracing.EnsureTracerProvider(tracerProvider).Tracer("outbound_emailer_job")).StartSpan(ctx)
-	defer span.End()
-
-	analyticsEventReporter, err := analyticsconfig.ProvideEventReporter(&cfg.Analytics, logger, tracerProvider)
-	if err != nil {
-		return observability.PrepareAndLogError(err, logger, span, "setting up customer data collector")
-	}
-
-	defer analyticsEventReporter.Close()
-
-	dbConnectionContext, cancel := context.WithTimeout(ctx, 15*time.Second)
-	dataManager, err := postgres.ProvideDatabaseClient(dbConnectionContext, logger, tracerProvider, &cfg.Database)
-	if err != nil {
-		cancel()
-		return observability.PrepareAndLogError(err, logger, span, "establishing database connection")
-	}
-
-	cancel()
-	defer dataManager.Close()
-
-	emailer, err := emailconfig.ProvideEmailer(&cfg.Email, logger, tracerProvider, otelhttp.DefaultClient)
-	if err != nil {
-		return observability.PrepareAndLogError(err, logger, span, "configuring outbound emailer")
-	}
-
 	var msg MessagePublishedData
 	if err = e.DataAs(&msg); err != nil {
 		return fmt.Errorf("event.DataAs: %w", err)
@@ -104,90 +69,8 @@ func SendEmail(ctx context.Context, e event.Event) error {
 	var emailDeliveryRequest email.DeliveryRequest
 	if err = json.Unmarshal(msg.Message.Data, &emailDeliveryRequest); err != nil {
 		logger = logger.WithValue("raw_data", msg.Message.Data)
-		return observability.PrepareAndLogError(err, logger, span, "unmarshalling delivery request message")
+		return observability.PrepareAndLogError(err, logger, nil, "unmarshalling delivery request message")
 	}
 
-	logger = logger.WithValue("template", emailDeliveryRequest.Template)
-
-	user, err := dataManager.GetUser(ctx, emailDeliveryRequest.UserID)
-	if err != nil {
-		return observability.PrepareAndLogError(err, logger, span, "getting user")
-	}
-
-	var (
-		mail                   *email.OutboundEmailMessage
-		shouldSkipIfUnverified = true
-		emailType              string
-	)
-
-	switch emailDeliveryRequest.Template {
-	case email.TemplateTypeInvite:
-		if emailDeliveryRequest.Invitation == nil {
-			return observability.PrepareAndLogError(err, logger, span, "missing household invitation")
-		}
-
-		mail, err = email.BuildInviteMemberEmail(emailDeliveryRequest.Invitation, envCfg)
-		if err != nil {
-			return observability.PrepareAndLogError(err, logger, span, "building email message")
-		}
-		emailType = "invite"
-	case email.TemplateTypeUsernameReminder:
-		mail, err = email.BuildUsernameReminderEmail(user, envCfg)
-		if err != nil {
-			return observability.PrepareAndLogError(err, logger, span, "building username reminder email")
-		}
-		emailType = "username reminder"
-	case email.TemplateTypePasswordResetTokenCreated:
-		if emailDeliveryRequest.PasswordResetToken == nil {
-			return observability.PrepareAndLogError(err, logger, span, "missing password reset token")
-		}
-
-		mail, err = email.BuildGeneratedPasswordResetTokenEmail(user, emailDeliveryRequest.PasswordResetToken, envCfg)
-		if err != nil {
-			return observability.PrepareAndLogError(err, logger, span, "building password reset token created email")
-		}
-		emailType = "password reset token"
-	case email.TemplateTypePasswordReset:
-		mail, err = email.BuildPasswordChangedEmail(user, envCfg)
-		if err != nil {
-			return observability.PrepareAndLogError(err, logger, span, "building password reset token email")
-		}
-		emailType = "password reset token"
-	case email.TemplateTypePasswordResetTokenRedeemed:
-		mail, err = email.BuildPasswordResetTokenRedeemedEmail(user, envCfg)
-		if err != nil {
-			return observability.PrepareAndLogError(err, logger, span, "building password reset token redemption email")
-		}
-		emailType = "password reset token redemption"
-	case email.TemplateTypeMealPlanCreated:
-		mail, err = email.BuildMealPlanCreatedEmail(user, emailDeliveryRequest.MealPlan, envCfg)
-		if err != nil {
-			return observability.PrepareAndLogError(err, logger, span, "building meal plan created email")
-		}
-		emailType = "meal plan created"
-	case email.TemplateTypeVerifyEmailAddress:
-		shouldSkipIfUnverified = false
-		mail, err = email.BuildVerifyEmailAddressEmail(user, emailDeliveryRequest.EmailVerificationToken, envCfg)
-		if err != nil {
-			return observability.PrepareAndLogError(err, logger, span, "building address verification email")
-		}
-		emailType = "email address verification"
-	}
-
-	if shouldSkipIfUnverified && user.EmailAddressVerifiedAt == nil {
-		logger.Info("user email address not verified, skipping email delivery")
-		return nil
-	}
-
-	logger.Info("sending email")
-
-	if err = emailer.SendEmail(ctx, mail); err != nil {
-		observability.AcknowledgeError(err, logger, span, "sending %s email", emailType)
-	}
-
-	if err = analyticsEventReporter.EventOccurred(ctx, email.SentEventType, emailDeliveryRequest.UserID, emailDeliveryRequest.TemplateParams); err != nil {
-		observability.AcknowledgeError(err, logger, span, "notifying customer data platform")
-	}
-
-	return nil
+	return logic.HandleSendEmailRequest(ctx, logger, tracerProvider, cfg, &emailDeliveryRequest)
 }
