+++ backend/cmd/services/api/http/main.go
@@ -80,6 +80,7 @@ func main() {
 	// only allow initialization to take so long.
 	buildCtx, cancel := context.WithTimeout(rootCtx, cfg.Server.StartupDeadline)
 
+	log.Println("building server")
 	// build our server struct.
 	srv, err := build.Build(buildCtx, cfg)
 	if err != nil {
@@ -97,6 +98,7 @@ func main() {
 		syscall.SIGTERM,
 	)
 
+	log.Println("serving")
 	// Run server
 	go srv.Serve()
 
@@ -111,6 +113,7 @@ func main() {
 	cancelCtx, cancelShutdown := context.WithTimeout(rootCtx, 10*time.Second)
 	defer cancelShutdown()
 
+	log.Println("shutting down")
 	// Gracefully shutdown the server by waiting on existing requests (except websockets).
 	if err = srv.Shutdown(cancelCtx); err != nil {
 		panic(err)
