+++ backend/environments/local/helm/charts/api-server/templates/tests/test-connection.yaml
@@ -1,0 +1,15 @@
+apiVersion: v1
+kind: Pod
+metadata:
+  name: "{{ include "api-server.fullname" . }}-test-connection"
+  labels:
+    {{- include "api-server.labels" . | nindent 4 }}
+  annotations:
+    "helm.sh/hook": test
+spec:
+  containers:
+    - name: wget
+      image: busybox
+      command: ['wget']
+      args: ['{{ include "api-server.fullname" . }}:{{ .Values.service.port }}']
+  restartPolicy: Never
