+++ backend/skaffold.yaml
@@ -1,0 +1,151 @@
+apiVersion: skaffold/v4beta11
+kind: Config
+
+metadata:
+  name: dinner-done-better-backend
+
+profiles:
+  - name: localdev
+
+    build:
+      tagPolicy:
+        sha256: {}
+      artifacts:
+        - image: dinner-done-better-api-server
+          context: .
+          ko:
+            dir: .
+            env:
+              - CGO_ENABLED=1
+            flags:
+              - -trimpath
+            ldflags:
+              - -s -w
+              - -extldflags "-static"
+            main: ./cmd/services/api/http
+            fromImage: golang:1.22-bullseye
+
+#        - image: ddb_job_db_cleaner
+#          context: .
+#          ko:
+#            dir: .
+#            env:
+#              - CGO_ENABLED=1
+#            flags:
+#              - -trimpath
+#            ldflags:
+#              - -s -w
+#              - -extldflags "-static"
+#            main: ./cmd/jobs/db_cleaner
+#            fromImage: gcr.io/distroless/static:nonroot
+#
+#        - image: ddb_job_email_prober
+#          context: .
+#          ko:
+#            dir: .
+#            env:
+#              - CGO_ENABLED=1
+#            flags:
+#              - -trimpath
+#            ldflags:
+#              - -s -w
+#              - -extldflags "-static"
+#            main: ./cmd/jobs/email_prober
+#            fromImage: gcr.io/distroless/static:nonroot
+#
+#        - image: ddb_job_meal_plan_finalizer
+#          context: .
+#          ko:
+#            dir: .
+#            env:
+#              - CGO_ENABLED=1
+#            flags:
+#              - -trimpath
+#            ldflags:
+#              - -s -w
+#              - -extldflags "-static"
+#            main: ./cmd/jobs/meal_plan_finalizer
+#            fromImage: gcr.io/distroless/static:nonroot
+#
+#        - image: ddb_job_meal_plan_grocery_list_initializer
+#          context: .
+#          ko:
+#            dir: .
+#            env:
+#              - CGO_ENABLED=1
+#            flags:
+#              - -trimpath
+#            ldflags:
+#              - -s -w
+#              - -extldflags "-static"
+#            main: ./cmd/jobs/meal_plan_grocery_list_initializer
+#            fromImage: gcr.io/distroless/static:nonroot
+#
+#        - image: ddb_job_meal_plan_task_creator
+#          context: .
+#          ko:
+#            dir: .
+#            env:
+#              - CGO_ENABLED=1
+#            flags:
+#              - -trimpath
+#            ldflags:
+#              - -s -w
+#              - -extldflags "-static"
+#            main: ./cmd/jobs/meal_plan_task_creator
+#            fromImage: gcr.io/distroless/static:nonroot
+#
+#        - image: ddb_job_search_data_index_scheduler
+#          context: .
+#          ko:
+#            dir: .
+#            env:
+#              - CGO_ENABLED=1
+#            flags:
+#              - -trimpath
+#            ldflags:
+#              - -s -w
+#              - -extldflags "-static"
+#            main: ./cmd/jobs/search_data_index_scheduler
+#            fromImage: gcr.io/distroless/static:nonroot
+
+        # TODO: cmd/functions
+
+    deploy:
+      kubeContext: docker-desktop
+      helm:
+        releases:
+          - name: postgres
+            namespace: localdev
+            createNamespace: true
+            remoteChart: oci://registry-1.docker.io/bitnamicharts/postgresql
+            version: 16.1.2
+            overrides:
+              global:
+                postgresql:
+                  auth:
+                    username: dbuser
+                    password: hunter2
+                    database: dinner-done-better
+          - name: redis
+            namespace: localdev
+            createNamespace: true
+            remoteChart: oci://registry-1.docker.io/bitnamicharts/redis
+            version: 20.2.1
+            overrides:
+              architecture: standalone
+
+          - name: dinnerdonebetter
+            namespace: localdev
+            createNamespace: true
+            chartPath: environments/local/helm/charts/api-server
+
+    portForward:
+      - resourceType: Service
+        namespace: localdev
+        resourceName: postgres-postgresql
+        port: 5423
+        localPort: 5434
+        address: 127.0.0.1
+
+  - name: dev
